move PlayerGameSummary and PlayerHistory to a new http.player package
