+++ game-service-server/src/main/scala/com/andy327/server/http/player/PlayerHistory.scala
@@ -1,4 +1,4 @@
-package com.andy327.server.http.auth
+package com.andy327.server.http.player
 
 import java.time.Instant
